pyproject.toml
[project]
name = "py-swisseph-natal-api"
version = "0.1.0"
description = "An AGPL compliant open source natal api powered by pyswisseph"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datetime>=5.5",
    "fastapi>=0.115.12",
    "pydantic>=2.11.4",
    "pyswisseph>=2.10.3.2",
    "pytz>=2025.2",
    "uvicorn>=0.34.2",
]
